在modules包下新建asset-management包，将代码挪过来
