+++ pom.xml
@@ -347,6 +347,18 @@
 
             <dependency>
                 <groupId>org.ruoyi</groupId>
+                <artifactId>ruoyi-asset-management-api</artifactId>
+                <version>${revision}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.ruoyi</groupId>
+                <artifactId>ruoyi-asset-management</artifactId>
+                <version>${revision}</version>
+            </dependency>
+
+            <dependency>
+                <groupId>org.ruoyi</groupId>
                 <artifactId>ruoyi-generator</artifactId>
                 <version>${revision}</version>
             </dependency>
