+++ ruoyi-admin/pom.xml
@@ -57,6 +57,11 @@
             <artifactId>ruoyi-generator</artifactId>
         </dependency>
 
+        <dependency>
+            <groupId>org.ruoyi</groupId>
+            <artifactId>ruoyi-asset-management</artifactId>
+        </dependency>
+
     </dependencies>
 
     <build>
