--- ruoyi-modules-api/ruoyi-system-api/src/main/java/org/ruoyi/system/mapper/MinUsagePeriodMapper.java
@@ -1,17 +1,0 @@
-package org.ruoyi.system.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.ruoyi.core.mapper.BaseMapperPlus;
-import org.ruoyi.system.domain.MinUsagePeriod;
-import org.ruoyi.system.domain.vo.MinUsagePeriodVo;
-
-/**
- * 最低使用年限表 数据层
- *
- * @author cass
- * @date 2025-09-24
- */
-@Mapper
-public interface MinUsagePeriodMapper extends BaseMapperPlus<MinUsagePeriod, MinUsagePeriodVo> {
-
-}
