+++ ruoyi-modules/ruoyi-asset-management/pom.xml
@@ -4,75 +4,82 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
     <parent>
-        <artifactId>ruoyi-ai</artifactId>
         <groupId>org.ruoyi</groupId>
+        <artifactId>ruoyi-modules</artifactId>
         <version>${revision}</version>
         <relativePath>../pom.xml</relativePath>
     </parent>
 
-    <artifactId>ruoyi-modules-api</artifactId>
-    <packaging>pom</packaging>
+    <artifactId>ruoyi-asset-management</artifactId>
 
-    <modules>
-        <module>ruoyi-chat-api</module>
-        <module>ruoyi-knowledge-api</module>
-        <module>ruoyi-system-api</module>
-        <module>ruoyi-asset-management-api</module>
-    </modules>
-
-    <properties>
-        <maven.compiler.source>17</maven.compiler.source>
-        <maven.compiler.target>17</maven.compiler.target>
-        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    </properties>
+    <description>
+        资产管理模块
+    </description>
 
     <dependencies>
-        <!-- 系统核心模块 -->
+        <!-- 通用工具-->
         <dependency>
             <groupId>org.ruoyi</groupId>
             <artifactId>ruoyi-common-core</artifactId>
         </dependency>
 
-        <!-- mybaits基础模块 -->
         <dependency>
             <groupId>org.ruoyi</groupId>
-            <artifactId>ruoyi-common-mybatis</artifactId>
+            <artifactId>ruoyi-common-doc</artifactId>
         </dependency>
 
-        <!-- 脱敏模块 -->
+        <!-- OSS功能模块 -->
         <dependency>
             <groupId>org.ruoyi</groupId>
-            <artifactId>ruoyi-common-sensitive</artifactId>
+            <artifactId>ruoyi-common-oss</artifactId>
         </dependency>
 
-        <!-- excel模块-->
+        <!-- excel-->
         <dependency>
             <groupId>org.ruoyi</groupId>
             <artifactId>ruoyi-common-excel</artifactId>
         </dependency>
 
-        <!-- 租户基础模块 -->
+        <!-- SMS功能模块 -->
         <dependency>
             <groupId>org.ruoyi</groupId>
-            <artifactId>ruoyi-common-tenant</artifactId>
+            <artifactId>ruoyi-common-sms</artifactId>
         </dependency>
 
-        <!-- 字段翻译基础模块 -->
         <dependency>
             <groupId>org.ruoyi</groupId>
-            <artifactId>ruoyi-common-translation</artifactId>
+            <artifactId>ruoyi-common-mail</artifactId>
         </dependency>
 
-        <!-- 系统日志模块 -->
+        <!-- 资产管理API模块 -->
         <dependency>
             <groupId>org.ruoyi</groupId>
-            <artifactId>ruoyi-common-log</artifactId>
+            <artifactId>ruoyi-asset-management-api</artifactId>
         </dependency>
 
-        <!-- 对象存储模块 -->
+        <!-- Apache Commons IO -->
         <dependency>
-            <groupId>org.ruoyi</groupId>
-            <artifactId>ruoyi-common-oss</artifactId>
+            <groupId>commons-io</groupId>
+            <artifactId>commons-io</artifactId>
+        </dependency>
+
+        <!-- 测试依赖 -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-test</artifactId>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter</artifactId>
+            <scope>test</scope>
+        </dependency>
+
+        <dependency>
+            <groupId>org.springframework</groupId>
+            <artifactId>spring-test</artifactId>
+            <scope>test</scope>
         </dependency>
 
         <dependency>
@@ -80,7 +87,7 @@
             <artifactId>mysql-connector-java</artifactId>
         </dependency>
 
-
     </dependencies>
 
 </project>
+
