+++ ruoyi-modules/ruoyi-system/pom.xml
@@ -68,6 +68,11 @@
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
         </dependency>
+        <dependency>
+            <groupId>org.ruoyi</groupId>
+            <artifactId>ruoyi-asset-management-api</artifactId>
+            <scope>test</scope>
+        </dependency>
 
     </dependencies>
