+++ ruoyi-modules/ruoyi-system/src/test/resources/application-test.yml
@@ -1,5 +1,6 @@
 # 测试环境配置
 spring:
+  # 数据源配置
   datasource:
     type: com.zaxxer.hikari.HikariDataSource
     dynamic:
@@ -31,27 +32,45 @@ spring:
       database: 0
       timeout: 10S
 
-# MyBatis Plus配置
-mybatis-plus:
-  mapper-locations: classpath*:mapper/**/*Mapper.xml
-  type-aliases-package: org.ruoyi.**.domain
-  configuration:
-    map-underscore-to-camel-case: true
-    cache-enabled: false
-    call-setters-on-nulls: true
-    jdbc-type-for-null: 'null'
-  global-config:
-    db-config:
-      id-type: AUTO
-      logic-delete-field: delFlag
-      logic-delete-value: 2
-      logic-not-delete-value: 0
+  # MyBatis Plus配置
+  mybatis-plus:
+    # 搜索指定包别名
+    typeAliasesPackage: org.ruoyi.**.domain
+    # 配置mapper的扫描，找到所有的mapper.xml映射文件
+    mapperLocations: classpath*:mapper/**/*Mapper.xml
+    # 加载全局的配置文件
+    configLocation: classpath:mybatis/mybatis-config.xml
+    configuration:
+      # 自动驼峰命名规则（camel case）映射
+      mapUnderscoreToCamelCase: true
+      # MyBatis 自动映射策略
+      # NONE：不启用 PARTIAL：只对非嵌套 resultMap 自动映射 FULL：对所有 resultMap 自动映射
+      autoMappingBehavior: PARTIAL
+      # MyBatis 自动映射时未知列或未知属性处理策
+      # NONE：不做处理 WARNING：打印相关警告 FAILING：抛出异常和详细信息
+      autoMappingUnknownColumnBehavior: NONE
+      # 更详细的日志输出 会有性能损耗 org.apache.ibatis.logging.stdout.StdOutImpl
+      # 关闭日志记录 (可单纯使用 p6spy 分析) org.apache.ibatis.logging.nologging.NoLoggingImpl
+      # 默认日志输出 org.apache.ibatis.logging.slf4j.Slf4jImpl
+      logImpl: org.apache.ibatis.logging.nologging.NoLoggingImpl
+    global-config:
+      # 是否打印 Logo banner
+      banner: true
+      dbConfig:
+        # 主键类型
+        # AUTO 自增 NONE 空 INPUT 用户输入 ASSIGN_ID 雪花 ASSIGN_UUID 唯一 UUID
+        idType: ASSIGN_ID
+        # 逻辑删除配置
+        logicDeleteField: delFlag
+        logicDeleteValue: 2
+        logicNotDeleteValue: 0
+        insertStrategy: NOT_NULL
+        updateStrategy: NOT_NULL
+        where-strategy: NOT_NULL
 
 # 日志配置
 logging:
   level:
-    org.ruoyi: DEBUG
-    org.springframework: WARN
-    com.baomidou.mybatisplus: WARN
-  pattern:
-    console: '%d{yyyy-MM-dd HH:mm:ss.SSS} [%thread] %-5level %logger{50} - %msg%n'
+    org.ruoyi: debug
+    org.springframework: warn
+    com.alibaba.excel: debug
